requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Document processing
langchain>=0.0.200
langchain-community>=0.0.10
pypdf2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Vector database and search
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Evaluation and metrics
evaluate>=0.4.0
sacrebleu>=2.0.0
rouge-score>=0.1.2
nltk>=3.8.1

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
